Move TxContext into the execution layer given it is "internal" to that component
