+++ crates/sui-types/src/base_types.rs
@@ -26,18 +26,15 @@ use crate::dynamic_field::DynamicFieldType;
 use crate::dynamic_field::{DYNAMIC_FIELD_FIELD_STRUCT_NAME, DYNAMIC_FIELD_MODULE_NAME};
 use crate::effects::TransactionEffects;
 use crate::effects::TransactionEffectsAPI;
-use crate::epoch_data::EpochData;
-use crate::error::ExecutionErrorKind;
 use crate::error::SuiError;
 use crate::error::SuiErrorKind;
-use crate::error::{ExecutionError, SuiResult};
+use crate::error::SuiResult;
 use crate::gas_coin::GAS;
 use crate::gas_coin::GasCoin;
 use crate::governance::STAKED_SUI_STRUCT_NAME;
 use crate::governance::STAKING_POOL_MODULE_NAME;
 use crate::governance::StakedSui;
 use crate::id::RESOLVED_SUI_ID;
-use crate::messages_checkpoint::CheckpointTimestamp;
 use crate::multisig::MultiSigPublicKey;
 use crate::object::{Object, Owner};
 use crate::parse_sui_struct_tag;
@@ -77,10 +74,9 @@ use serde_with::serde_as;
 use shared_crypto::intent::HashingIntentScope;
 use std::borrow::Cow;
 use std::cmp::max;
-use std::convert::{TryFrom, TryInto};
+use std::convert::TryFrom;
 use std::fmt;
 use std::str::FromStr;
-use sui_protocol_config::ProtocolConfig;
 
 #[cfg(test)]
 #[path = "unit_tests/base_types_tests.rs"]
@@ -1176,68 +1172,8 @@ pub fn url_layout() -> A::MoveStructLayout {
     }
 }
 
-// The Rust representation of the Move `TxContext`.
-// This struct must be kept in sync with the Move `TxContext` definition.
-// Moving forward we are going to zero all fields of the Move `TxContext`
-// and use native functions to retrieve info about the transaction.
-// However we cannot remove the Move type and so this struct is going to
-// be the Rust equivalent to the Move `TxContext` for legacy usages.
-//
-// `TxContext` in Rust (see below) is going to be purely used in Rust and can
-// evolve as needed without worrying any compatibility with Move.
-#[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Eq)]
-pub struct MoveLegacyTxContext {
-    // Signer/sender of the transaction
-    sender: AccountAddress,
-    // Digest of the current transaction
-    digest: Vec<u8>,
-    // The current epoch number
-    epoch: EpochId,
-    // Timestamp that the epoch started at
-    epoch_timestamp_ms: CheckpointTimestamp,
-    // Number of `ObjectID`'s generated during execution of the current transaction
-    ids_created: u64,
-}
-
-impl From<&TxContext> for MoveLegacyTxContext {
-    fn from(tx_context: &TxContext) -> Self {
-        Self {
-            sender: tx_context.sender,
-            digest: tx_context.digest.clone(),
-            epoch: tx_context.epoch,
-            epoch_timestamp_ms: tx_context.epoch_timestamp_ms,
-            ids_created: tx_context.ids_created,
-        }
-    }
-}
-
-// Information about the transaction context.
-// This struct is not related to Move and can evolve as needed/required.
-#[derive(Clone, Debug, Deserialize, Serialize, PartialEq, Eq)]
-pub struct TxContext {
-    /// Sender of the transaction
-    sender: AccountAddress,
-    /// Digest of the current transaction
-    digest: Vec<u8>,
-    /// The current epoch number
-    epoch: EpochId,
-    /// Timestamp that the epoch started at
-    epoch_timestamp_ms: CheckpointTimestamp,
-    /// Number of `ObjectID`'s generated during execution of the current transaction
-    ids_created: u64,
-    // Reference gas price
-    rgp: u64,
-    // gas price passed to transaction as input
-    gas_price: u64,
-    // gas budget passed to transaction as input
-    gas_budget: u64,
-    // address of the sponsor if any
-    sponsor: Option<AccountAddress>,
-    // whether the `TxContext` is native or not
-    // (TODO: once we version execution we could drop this field)
-    is_native: bool,
-}
-
+// Whether a given argument (a `SignatureToken`) is a Move `TxContext` or
+// a reference (mutable or immutable) to it.
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum TxContextKind {
     // No TxContext
@@ -1248,57 +1184,9 @@ pub enum TxContextKind {
     Immutable,
 }
 
-impl TxContext {
-    pub fn new(
-        sender: &SuiAddress,
-        digest: &TransactionDigest,
-        epoch_data: &EpochData,
-        rgp: u64,
-        gas_price: u64,
-        gas_budget: u64,
-        sponsor: Option<SuiAddress>,
-        protocol_config: &ProtocolConfig,
-    ) -> Self {
-        Self::new_from_components(
-            sender,
-            digest,
-            &epoch_data.epoch_id(),
-            epoch_data.epoch_start_timestamp(),
-            rgp,
-            gas_price,
-            gas_budget,
-            sponsor,
-            protocol_config,
-        )
-    }
-
-    pub fn new_from_components(
-        sender: &SuiAddress,
-        digest: &TransactionDigest,
-        epoch_id: &EpochId,
-        epoch_timestamp_ms: u64,
-        rgp: u64,
-        gas_price: u64,
-        gas_budget: u64,
-        sponsor: Option<SuiAddress>,
-        protocol_config: &ProtocolConfig,
-    ) -> Self {
-        Self {
-            sender: AccountAddress::new(sender.0),
-            digest: digest.into_inner().to_vec(),
-            epoch: *epoch_id,
-            epoch_timestamp_ms,
-            ids_created: 0,
-            rgp,
-            gas_price,
-            gas_budget,
-            sponsor: sponsor.map(|s| s.into()),
-            is_native: protocol_config.move_native_context(),
-        }
-    }
-
+impl TxContextKind {
     /// Returns whether the type signature is &mut TxContext, &TxContext, or none of the above.
-    pub fn kind(view: &CompiledModule, s: &SignatureToken) -> TxContextKind {
+    pub fn derive(view: &CompiledModule, s: &SignatureToken) -> Self {
         use SignatureToken as S;
         let (kind, s) = match s {
             S::MutableReference(s) => (TxContextKind::Mutable, s),
@@ -1315,131 +1203,6 @@ impl TxContext {
         } else {
             TxContextKind::None
         }
-    }
-
-    pub fn type_() -> StructTag {
-        StructTag {
-            address: SUI_FRAMEWORK_ADDRESS,
-            module: TX_CONTEXT_MODULE_NAME.to_owned(),
-            name: TX_CONTEXT_STRUCT_NAME.to_owned(),
-            type_params: vec![],
-        }
-    }
-
-    pub fn epoch(&self) -> EpochId {
-        self.epoch
-    }
-
-    pub fn sender(&self) -> SuiAddress {
-        self.sender.into()
-    }
-
-    pub fn epoch_timestamp_ms(&self) -> u64 {
-        self.epoch_timestamp_ms
-    }
-
-    /// Return the transaction digest, to include in new objects
-    pub fn digest(&self) -> TransactionDigest {
-        TransactionDigest::new(self.digest.clone().try_into().unwrap())
-    }
-
-    pub fn sponsor(&self) -> Option<SuiAddress> {
-        self.sponsor.map(SuiAddress::from)
-    }
-
-    pub fn rgp(&self) -> u64 {
-        self.rgp
-    }
-
-    pub fn gas_price(&self) -> u64 {
-        self.gas_price
-    }
-
-    pub fn gas_budget(&self) -> u64 {
-        self.gas_budget
-    }
-
-    pub fn ids_created(&self) -> u64 {
-        self.ids_created
-    }
-
-    /// Derive a globally unique object ID by hashing self.digest | self.ids_created
-    pub fn fresh_id(&mut self) -> ObjectID {
-        let id = ObjectID::derive_id(self.digest(), self.ids_created);
-
-        self.ids_created += 1;
-        id
-    }
-
-    pub fn to_bcs_legacy_context(&self) -> Vec<u8> {
-        let move_context: MoveLegacyTxContext = if self.is_native {
-            let tx_context = &TxContext {
-                sender: AccountAddress::ZERO,
-                digest: self.digest.clone(),
-                epoch: 0,
-                epoch_timestamp_ms: 0,
-                ids_created: 0,
-                rgp: 0,
-                gas_price: 0,
-                gas_budget: 0,
-                sponsor: None,
-                is_native: true,
-            };
-            tx_context.into()
-        } else {
-            self.into()
-        };
-        bcs::to_bytes(&move_context).unwrap()
-    }
-
-    pub fn to_vec(&self) -> Vec<u8> {
-        bcs::to_bytes(&self).unwrap()
-    }
-
-    /// Updates state of the context instance. It's intended to use
-    /// when mutable context is passed over some boundary via
-    /// serialize/deserialize and this is the reason why this method
-    /// consumes the other context..
-    pub fn update_state(&mut self, other: MoveLegacyTxContext) -> Result<(), ExecutionError> {
-        if !self.is_native {
-            if self.sender != other.sender
-                || self.digest != other.digest
-                || other.ids_created < self.ids_created
-            {
-                return Err(ExecutionError::new_with_source(
-                    ExecutionErrorKind::InvariantViolation,
-                    "Immutable fields for TxContext changed",
-                ));
-            }
-            self.ids_created = other.ids_created;
-        }
-        Ok(())
-    }
-
-    //
-    // Move test only API
-    //
-    pub fn replace(
-        &mut self,
-        sender: AccountAddress,
-        tx_hash: Vec<u8>,
-        epoch: u64,
-        epoch_timestamp_ms: u64,
-        ids_created: u64,
-        rgp: u64,
-        gas_price: u64,
-        gas_budget: u64,
-        sponsor: Option<AccountAddress>,
-    ) {
-        self.sender = sender;
-        self.digest = tx_hash;
-        self.epoch = epoch;
-        self.epoch_timestamp_ms = epoch_timestamp_ms;
-        self.ids_created = ids_created;
-        self.rgp = rgp;
-        self.gas_price = gas_price;
-        self.gas_budget = gas_budget;
-        self.sponsor = sponsor;
     }
 }
