+++ sui-execution/latest/sui-adapter/src/static_programmable_transactions/loading/translate.rs
@@ -11,8 +11,9 @@ use crate::{
     },
 };
 use move_core_types::{account_address::AccountAddress, language_storage::StructTag, u256::U256};
+use sui_execution_types::TxContext;
 use sui_types::{
-    base_types::{ObjectID, TxContext},
+    base_types::ObjectID,
     error::ExecutionError,
     object::Owner,
     transaction::{self as P, CallArg, FundsWithdrawalArg, ObjectArg, SharedObjectMutability},
