+++ sui-execution/latest/sui-adapter/src/static_programmable_transactions/mod.rs
@@ -17,10 +17,11 @@ use crate::{
 use move_trace_format::format::MoveTraceBuilder;
 use move_vm_runtime::move_vm::MoveVM;
 use std::{cell::RefCell, rc::Rc, sync::Arc};
+use sui_execution_types::TxContext;
 use sui_protocol_config::ProtocolConfig;
 use sui_types::{
-    base_types::TxContext, error::ExecutionError, execution::ResultWithTimings,
-    metrics::LimitsMetrics, storage::BackingPackageStore, transaction::ProgrammableTransaction,
+    error::ExecutionError, execution::ResultWithTimings, metrics::LimitsMetrics,
+    storage::BackingPackageStore, transaction::ProgrammableTransaction,
 };
 
 // TODO we might replace this with a new one
