+++ sui-execution/latest/sui-move-natives/Cargo.toml
@@ -28,5 +28,6 @@ move-vm-runtime = { path = "../../../external-crates/move/crates/move-vm-runtime
 move-vm-types = { path = "../../../external-crates/move/crates/move-vm-types" }
 
 sui-protocol-config.workspace = true
+sui-execution-types.workspace = true
 sui-types.workspace = true
 tracing.workspace = true
