+++ sui-execution/sui-execution-types/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "sui-execution-types"
+version = "0.1.0"
+authors = ["Mysten Labs <eng@mystenlabs.com>"]
+description = "Types internal to the Sui execution layer."
+license = "Apache-2.0"
+publish = false
+edition = "2024"
+
+[lints]
+workspace = true
+
+[dependencies]
+bcs.workspace = true
+serde.workspace = true
+
+move-core-types.workspace = true
+
+sui-protocol-config.workspace = true
+sui-types.workspace = true
