+++ sui-execution/v0/sui-adapter/Cargo.toml
@@ -33,6 +33,7 @@ move-vm-types = { path = "../../../external-crates/move/move-execution/v0/crates
 
 sui-macros.workspace = true
 sui-protocol-config.workspace = true
+sui-execution-types.workspace = true
 sui-types.workspace = true
 parking_lot.workspace = true
