+++ sui-execution/v0/sui-adapter/src/programmable_transactions/context.rs
@@ -32,6 +32,7 @@ mod checked {
     };
     use move_vm_runtime::{move_vm::MoveVM, session::Session};
     use move_vm_types::loaded_data::runtime_types::Type;
+    use sui_execution_types::TxContext;
     use sui_move_natives::object_runtime::{
         self, get_all_uids, max_event_error, ObjectRuntime, RuntimeResults,
     };
@@ -40,7 +41,7 @@ mod checked {
     use sui_types::storage::PackageObject;
     use sui_types::{
         balance::Balance,
-        base_types::{MoveObjectType, ObjectID, SequenceNumber, SuiAddress, TxContext},
+        base_types::{MoveObjectType, ObjectID, SequenceNumber, SuiAddress},
         coin::Coin,
         error::{command_argument_error, ExecutionError, ExecutionErrorKind},
         event::Event,
