+++ sui-execution/v1/sui-adapter/Cargo.toml
@@ -32,6 +32,7 @@ move-vm-types = { path = "../../../external-crates/move/move-execution/v1/crates
 
 sui-macros.workspace = true
 sui-protocol-config.workspace = true
+sui-execution-types.workspace = true
 sui-types.workspace = true
 parking_lot.workspace = true
