+++ sui-execution/v1/sui-verifier/Cargo.toml
@@ -17,4 +17,5 @@ move-abstract-stack.workspace = true
 
 move-bytecode-verifier = { path = "../../../external-crates/move/move-execution/v1/crates/move-bytecode-verifier", package = "move-bytecode-verifier-v1" }
 
+sui-execution-types.workspace = true
 sui-types.workspace = true
