+++ sui-execution/v1/sui-verifier/src/entry_points_verifier.rs
@@ -7,7 +7,7 @@ use move_binary_format::{
 };
 use move_bytecode_utils::format_signature_token;
 use sui_types::{
-    base_types::{TxContext, TxContextKind, TX_CONTEXT_MODULE_NAME, TX_CONTEXT_STRUCT_NAME},
+    base_types::{TxContextKind, TX_CONTEXT_MODULE_NAME, TX_CONTEXT_STRUCT_NAME},
     clock::Clock,
     error::ExecutionError,
     is_object, is_object_vector, is_primitive,
@@ -149,7 +149,7 @@ fn verify_init_function(module: &CompiledModule, fdef: &FunctionDefinition) -> R
     // then the first parameter must be of a one-time witness type and must be passed by value. This
     // is checked by the verifier for pass one-time witness value (one_time_witness_verifier) -
     // please see the description of this pass for additional details.
-    if TxContext::kind(module, &parameters[parameters.len() - 1]) != TxContextKind::None {
+    if TxContextKind::derive(module, &parameters[parameters.len() - 1]) != TxContextKind::None {
         Ok(())
     } else {
         Err(format!(
@@ -173,7 +173,7 @@ fn verify_entry_function_impl(
     let params = module.signature_at(handle.parameters);
 
     let all_non_ctx_params = match params.0.last() {
-        Some(last_param) if TxContext::kind(module, last_param) != TxContextKind::None => {
+        Some(last_param) if TxContextKind::derive(module, last_param) != TxContextKind::None => {
             &params.0[0..params.0.len() - 1]
         }
         _ => &params.0,
