+++ sui-execution/v2/sui-adapter/Cargo.toml
@@ -32,6 +32,7 @@ move-vm-types = { path = "../../../external-crates/move/move-execution/v2/crates
 
 sui-macros.workspace = true
 sui-protocol-config.workspace = true
+sui-execution-types.workspace = true
 sui-types.workspace = true
 parking_lot.workspace = true
