+++ sui-execution/v2/sui-verifier/Cargo.toml
@@ -17,5 +17,6 @@ move-abstract-stack.workspace = true
 
 move-bytecode-verifier = { path = "../../../external-crates/move/move-execution/v2/crates/move-bytecode-verifier", package = "move-bytecode-verifier-v2" }
 
+sui-execution-types.workspace = true
 sui-types.workspace = true
 sui-protocol-config.workspace = true
